Penggabungan seluruh kodingan UI/UX
