+++ lib/main.dart
@@ -1,12 +1,10 @@
 import 'package:flutter/material.dart';
 import 'pages/WelcomeScreen.dart';
-import 'package:flutter/rendering.dart';
-import 'package:google_fonts/google_fonts.dart';
-import 'package:route_wise/Pages/PengirimanBarang.dart';
 
 void main() {
   runApp(MyApp());
 }
+
 class MyApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
@@ -15,4 +13,3 @@ class MyApp extends StatelessWidget {
     );
   }
 }
-
